Renamed and merged some counter helpers
