+++ app/helpers/count_helper.rb
@@ -1,4 +1,4 @@
-module LanguageHelper
+module CountHelper
   def project_count_for_language
     Project.active.by_language(@language).count
   end
@@ -9,5 +9,15 @@ module LanguageHelper
 
   def user_count_for_language
     User.by_language(@language).count
+  end
+
+  def pull_request_count
+    return pull_request_count_for_language if @language
+    PullRequest.year(CURRENT_YEAR).count
+  end
+
+  def user_count
+    return user_count_for_language if @language
+    User.count
   end
 end
