--- app/helpers/pull_request_helper.rb
@@ -1,6 +1,0 @@
-module PullRequestHelper
-  def pull_request_count
-    return pull_request_count_for_language if @language
-    PullRequest.year(CURRENT_YEAR).count
-  end
-end
